Bug 1172264 - Switch MediaDecoder's mDuration represenation to a double. r=jww
UltraBlame original commit: 3ead3466f84a2f9b7f6b2ddd47ebd93b425b3691

+++ dom/media/MediaDecoder.h
@@ -944,9 +944,7 @@ public:
 protected:
 
   
-  
-  
-  int64_t mDuration;
+  double mDuration;
 
   
   bool mMediaSeekable;
